+++ lib/src/helpers/extensions.dart
@@ -1,0 +1,5 @@
+extension MyDateExtension on DateTime {
+  DateTime firstDayOfMonth(){
+    return DateTime(this.year, this.month);
+  }
+}